feat: restructure settings page into responsive grid layout
- Split settings into General and Appearance cards
- Added GeneralSettings and AppearanceSettings components
- Moved schema to a dedicated file
- Implemented side-by-side layout on desktop and stacked on mobile
